refactor: centralize admin status check and server-side data loading
